rename default and aboutus page
