--- src/pages/AboutUs/AboutUs.jsx
@@ -1,24 +1,0 @@
-import React from "react";
-import AboutusLinkPro from "./AboutusLinkPro";
-import AboutusWhyChoose from "./AboutusWhyChoose";
-import AboutUsWhatWeDo from "./AboutUsWhatWeDo";
-import AboutOurValues from "./AboutOurValues";
-import AboutTestimonials from "./AboutTestimonials";
-import DefaultBanner from "../../components/DefaultBanner.jsx";
-// import AboutMeetOurTeam from './AboutMeetOurTeam'
-
-function AboutUs() {
-  return (
-    <>
-      <DefaultBanner componentName="About Us" />
-      <AboutusLinkPro />
-      <AboutusWhyChoose />
-      <AboutUsWhatWeDo />
-      <AboutOurValues />
-      <AboutTestimonials />
-      {/* <AboutMeetOurTeam/> */}
-    </>
-  );
-}
-
-export default AboutUs;
