+++ src/pages/ContactUs/ContactUs.jsx
@@ -1,15 +1,14 @@
-import React from 'react'
-import GetInTouch from './GetInTouch'
-import DefaultBanner from '../../components/DefaultBanner'
+import React from "react";
+import GetInTouch from "./GetInTouch";
+import DefaultBanner from "../../components/DefBanner";
 
 function ContactUs() {
   return (
     <>
-    <DefaultBanner componentName="Contact Us"/>
-    <GetInTouch/>
-    
+      <DefaultBanner componentName="Contact Us" />
+      <GetInTouch />
     </>
-  )
+  );
 }
 
-export default ContactUs
+export default ContactUs;
