+++ src/pages/Services/Services.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import DefaultBanner from "../../components/DefaultBanner";
+import DefaultBanner from "../../components/DefBanner";
 import ServicesLinkProBest from "./serviceslinkprobest";
 import ServicesWhyChooseUs from "./serviceswhychooseus";
 import ServicesWhatWeDo from "./serviceswhatwedo";
@@ -8,9 +8,9 @@ function Services() {
   return (
     <>
       <DefaultBanner componentName="Services" />
-      <ServicesLinkProBest/>
-      <ServicesWhyChooseUs/>
-      <ServicesWhatWeDo/>
+      <ServicesLinkProBest />
+      <ServicesWhyChooseUs />
+      <ServicesWhatWeDo />
     </>
   );
 }
